Removed a - 1,  fixed 3 days of debugging :angry:
are you kidding me.
this is the epitomy of coding pain.
literally did so much work to create a whole new histogram function because just raw comparison was very surprisingly kinda dog at its job, but as it turns out, on a really high level I just shouldn't be subtracting the one in that particular place :sob:  :sob: :sob: